split detect into its own file, add `cg#*_check` convenience methods
